// packages/mpl-core-kit-lib/src/generated/types/baseOracle.ts
/**
 * This code was AUTOGENERATED using the codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  type Address,
  type Codec,
  combineCodec,
  type Decoder,
  type Encoder,
  getAddressDecoder,
  getAddressEncoder,
  getOptionDecoder,
  getOptionEncoder,
  getStructDecoder,
  getStructEncoder,
  type Option,
  type OptionOrNullable,
} from '@solana/kit'

import {
  type BaseExtraAccount,
  type BaseExtraAccountArgs,
  type BaseValidationResultsOffset,
  type BaseValidationResultsOffsetArgs,
  getBaseExtraAccountDecoder,
  getBaseExtraAccountEncoder,
  getBaseValidationResultsOffsetDecoder,
  getBaseValidationResultsOffsetEncoder,
} from './index.ts'

export type BaseOracle = {
  baseAddress: Address
  baseAddressConfig: Option<BaseExtraAccount>
  resultsOffset: BaseValidationResultsOffset
}

export type BaseOracleArgs = {
  baseAddress: Address
  baseAddressConfig: OptionOrNullable<BaseExtraAccountArgs>
  resultsOffset: BaseValidationResultsOffsetArgs
}

export function getBaseOracleCodec(): Codec<BaseOracleArgs, BaseOracle> {
  return combineCodec(getBaseOracleEncoder(), getBaseOracleDecoder())
}

export function getBaseOracleDecoder(): Decoder<BaseOracle> {
  return getStructDecoder([
    ['baseAddress', getAddressDecoder()],
    ['baseAddressConfig', getOptionDecoder(getBaseExtraAccountDecoder())],
    ['resultsOffset', getBaseValidationResultsOffsetDecoder()],
  ])
}

export function getBaseOracleEncoder(): Encoder<BaseOracleArgs> {
  return getStructEncoder([
    ['baseAddress', getAddressEncoder()],
    ['baseAddressConfig', getOptionEncoder(getBaseExtraAccountEncoder())],
    ['resultsOffset', getBaseValidationResultsOffsetEncoder()],
  ])
}