// packages/mpl-core-kit-lib/src/generated/types/externalRegistryRecord.ts
/**
 * This code was AUTOGENERATED using the codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  type Codec,
  combineCodec,
  type Decoder,
  type Encoder,
  getArrayDecoder,
  getArrayEncoder,
  getOptionDecoder,
  getOptionEncoder,
  getStructDecoder,
  getStructEncoder,
  getTupleDecoder,
  getTupleEncoder,
  getU64Decoder,
  getU64Encoder,
  type Option,
  type OptionOrNullable,
} from '@solana/kit'

import {
  type BasePluginAuthority,
  type BasePluginAuthorityArgs,
  type ExternalCheckResult,
  type ExternalCheckResultArgs,
  type ExternalPluginAdapterType,
  type ExternalPluginAdapterTypeArgs,
  getBasePluginAuthorityDecoder,
  getBasePluginAuthorityEncoder,
  getExternalCheckResultDecoder,
  getExternalCheckResultEncoder,
  getExternalPluginAdapterTypeDecoder,
  getExternalPluginAdapterTypeEncoder,
  getHookableLifecycleEventDecoder,
  getHookableLifecycleEventEncoder,
  type HookableLifecycleEvent,
  type HookableLifecycleEventArgs,
} from './index.ts'

export type ExternalRegistryRecord = {
  authority: BasePluginAuthority
  dataLen: Option<bigint>
  dataOffset: Option<bigint>
  lifecycleChecks: Option<Array<readonly [HookableLifecycleEvent, ExternalCheckResult]>>
  offset: bigint
  pluginType: ExternalPluginAdapterType
}

export type ExternalRegistryRecordArgs = {
  authority: BasePluginAuthorityArgs
  dataLen: OptionOrNullable<bigint | number>
  dataOffset: OptionOrNullable<bigint | number>
  lifecycleChecks: OptionOrNullable<Array<readonly [HookableLifecycleEventArgs, ExternalCheckResultArgs]>>
  offset: bigint | number
  pluginType: ExternalPluginAdapterTypeArgs
}

export function getExternalRegistryRecordCodec(): Codec<ExternalRegistryRecordArgs, ExternalRegistryRecord> {
  return combineCodec(getExternalRegistryRecordEncoder(), getExternalRegistryRecordDecoder())
}

export function getExternalRegistryRecordDecoder(): Decoder<ExternalRegistryRecord> {
  return getStructDecoder([
    ['pluginType', getExternalPluginAdapterTypeDecoder()],
    ['authority', getBasePluginAuthorityDecoder()],
    [
      'lifecycleChecks',
      getOptionDecoder(
        getArrayDecoder(getTupleDecoder([getHookableLifecycleEventDecoder(), getExternalCheckResultDecoder()])),
      ),
    ],
    ['offset', getU64Decoder()],
    ['dataOffset', getOptionDecoder(getU64Decoder())],
    ['dataLen', getOptionDecoder(getU64Decoder())],
  ])
}

export function getExternalRegistryRecordEncoder(): Encoder<ExternalRegistryRecordArgs> {
  return getStructEncoder([
    ['pluginType', getExternalPluginAdapterTypeEncoder()],
    ['authority', getBasePluginAuthorityEncoder()],
    [
      'lifecycleChecks',
      getOptionEncoder(
        getArrayEncoder(getTupleEncoder([getHookableLifecycleEventEncoder(), getExternalCheckResultEncoder()])),
      ),
    ],
    ['offset', getU64Encoder()],
    ['dataOffset', getOptionEncoder(getU64Encoder())],
    ['dataLen', getOptionEncoder(getU64Encoder())],
  ])
}